ragagent/backend/APP/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
ollama>=0.1.0
faiss-cpu>=1.7.4
cassandra-driver>=3.25.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
sentence-transformers>=2.2.0
transformers==4.30.0
torch>=2.0.0
langchain>=0.1.0
python-ollama>=0.1.0
langchain-community>=0.0.29
huggingface-hub==0.17.3
